Revert "Merge branch 'main' into refactor/clean-up-redundant-files-and-refine-logic"
This reverts commit e98c6f157d68f4a6bb04db0431596f6191bd6468, reversing
changes made to cdeab422788bf6132f65fb555829fbdb72daf456.
