--- data/batch/test2.csv
@@ -1,33 +1,0 @@
-text
-Allocating you S1
-Down all masts
-S9 has been identified
-solution is lagging TMA2
-Looking good team
-Just monitoring Sierra five
-can I get an updated solution
-S9 is at 180 ATB
-roger that
-"Stop, that's not correct"
-"Watchleader, check system"
-Sonar holds two new contacts. Merchant bearing 108 designated Sierra 1 and Merchant bearing 324 designated Sierra 2.
-All stations this is Track Manager. Please prioritise Sierra 1 as it appears substantially closer based upon its acoustic strength.
-Roger Track Manager; prioritising Sierra 1.
-"Roger, Track Manager. Raising periscope."
-"All stations, I confirm Sierra 1 is a merchant well within visual range. Getting range estimate now."
-Roger Optix. Sierra 1 within visual range. TMA please reduce range of Sierra 1 to around 10 kiloyards and limit speed to no more than 15 knots.
-Roger Track Manager. Sierra 1 range reduced to 10 kiloyards. Finding solution now.
-Thank you TMA.
-"Track Manager, sending through range estimate of Sierra 1 now. Estimated range 6.12 kiloyards with an ATB of 90 Port."
-"Thank you Optix. Fusing data now. TMA, Sierra 1 is now designated Foxtrot 1 with initial range and ATB estimates."
-Thank you Track Manager. Received in good order.
-"All stations, new solution for Foxtrot 1 entered."
-"Thank you TMA. Solution looks good. Please work on Sierra 2. Optix, is Sierra 2 within visual range."
-"Negative Track Manager. S2 beyond visual range, somewhere between 19 and 30 kiloyards."
-"All stations, entering initial solution for Sierra 2 now."
-Thank you TMA. Solution looks good.
-"Optix, this is Track Manager. Solution for Sierra 2 is beginning to lag. Recommend decreasing range somewhere between 19 and 27 kiloyards."
-"All stations, the tips of Sierra 2 are starting to emerge on the horizon. Entering new range estimate of 19 kiloyards."
-"Roger, thank you Optix. Update received. Reducing range of Foxtrot 2 to 19 kiloyards."
-"All stations, updating solution for Foxtrot 2 now."
-"Good work, TMA. Bearing rates now match for Foxtrot 2."
