+++ requirements.txt
@@ -1,3 +1,3 @@
 pandas==2.2.3
 streamlit==1.44.1
-matplotlib
+matplotlib==3.10.1
